ml-service/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Image processing
Pillow==10.2.0
numpy==1.26.4

# ML/Deep Learning
torch==2.2.0
torchvision==0.17.0

# HTTP client for downloading images/videos
httpx==0.26.0
aiohttp==3.9.3

# Video processing
opencv-python-headless==4.9.0.80

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
